renamed e2ee enabled devices to registered devices
this avoids confusion with e2ee enabled state for the plugin
