+++ app/products/e2ee/constants/database.ts
@@ -2,5 +2,5 @@
 // See LICENSE.txt for license information.
 
 export const E2EE_TABLES = {
-    E2EE_ENABLED_DEVICES: 'E2EEEnabledDevices',
+    E2EE_REGISTERED_DEVICES: 'E2EERegisteredDevices',
 };
