+++ app/products/e2ee/database/schema/index.ts
@@ -1,4 +1,4 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-export {default as E2EEEnabledDeviceSchema} from './e2ee_enabled_device';
+export {default as E2EERegisteredDeviceSchema} from './e2ee_registered_device';
